+++ honei/backends/cuda/sum.cu
@@ -39,7 +39,7 @@ extern "C" void cuda_sum_two_float(void * x, void * y, unsigned long size, unsig
     dim3 grid;
     dim3 block;
     block.x = blocksize;
-    grid.x = ceil(sqrt(size/(double)block.x));
+    grid.x = (unsigned)ceil(sqrt(size/(double)block.x));
     grid.y = grid.x;
     float * x_gpu((float *)x);
     float * y_gpu((float *)y);
